# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'HaloSitek Core API'
  description: 'OpenAPI metadata for HaloSitek Backoffice API.'
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000/api/v1'
    description: 'Local Development Server'
paths:
...
  /architects:
    get:
      tags:
        - Architects
      summary: 'List architects'
      description: 'Returns a list of architects for public browsing.'
      operationId: 1959934f63e0c7d99ea9d206a9ec85f0
      responses:
        200:
          description: 'Architect list retrieved successfully'
          content:
            application/json:
              schema: []
              example:
                success: true
                status_code: 200
                message: 'Architect list retrieved successfully'
                data:
                  -
                    id: 01HZX9M1F45M2Z6K7T9K7Y8QRA
                    name: 'Architect User'
                    email: architect@halositek.com
                    headline: 'Residential Specialist'
                meta:
                  current_page: 1
                  last_page: 1
                  per_page: 12
                  total: 1
                links:
                  first_page_url: 'http://localhost:8000/api/v1/architects?page=1'
                  last_page_url: 'http://localhost:8000/api/v1/architects?page=1'
                  next_page_url: null
                  prev_page_url: null
        404:
          $ref: '#/components/responses/NotFoundError'
        500:
          $ref: '#/components/responses/ServerError'
      security: []
  /architects/wishlist:
    get:
      tags:
        - Architects
      summary: 'List wishlist architects'
      description: "Returns the authenticated user's saved architect wishlist."
      operationId: 2a365608228314d6e7fea540e31572a0
      responses:
        200:
          description: 'Wishlist architects retrieved successfully'
          content:
            application/json:
              schema: []
              example:
                success: true
                status_code: 200
                message: 'Wishlist architects retrieved successfully'
                data:
                  -
                    id: 01HZX9M1F45M2Z6K7T9K7Y8QRA
                    name: 'Architect User'
                    email: architect@halositek.com
                    headline: 'Residential Specialist'
                meta:
                  current_page: 1
                  last_page: 1
                  per_page: 12
                  total: 1
                links:
                  first_page_url: 'http://localhost:8000/api/v1/architects/wishlist?page=1'
                  last_page_url: 'http://localhost:8000/api/v1/architects/wishlist?page=1'
                  next_page_url: null
                  prev_page_url: null
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          $ref: '#/components/responses/NotFoundError'
        500:
          $ref: '#/components/responses/ServerError'
      security:
        -
          BearerAuth: []
  '/architects/{userId}/save':
    post:
      tags:
        - Architects
      summary: 'Save architect to wishlist'
      description: "Saves a specific architect to the authenticated user's wishlist."
      operationId: 01aa4e3f3427e989cf2516c3dd89f86f
      parameters:
        -
          name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: 'Architect saved to wishlist successfully'
          content:
            application/json:
              schema: []
              example:
                success: true
                status_code: 200
                message: 'Architect saved to wishlist successfully'
                data: null
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          $ref: '#/components/responses/NotFoundError'
        409:
          $ref: '#/components/responses/ConflictError'
        500:
          $ref: '#/components/responses/ServerError'
      security:
        -
          BearerAuth: []
    delete:
      tags:
        - Architects
      summary: 'Remove architect from wishlist'
      description: "Removes a specific architect from the authenticated user's wishlist."
      operationId: ec27c9082ec363e91ed53ecf32a9805a
      parameters:
        -
          name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: 'Architect removed from wishlist successfully'
          content:
            application/json:
              schema: []
              example:
                success: true
                status_code: 200
                message: 'Architect removed from wishlist successfully'
                data: null
        401:
          $ref: '#/components/responses/UnauthorizedError'
        404:
          $ref: '#/components/responses/NotFoundError'
        500:
          $ref: '#/components/responses/ServerError'
      security:
        -
          BearerAuth: []
...
components:
  schemas:
...
    ApiError:
      properties:
        success:
          type: boolean
          example: false
        status_code:
          type: integer
          example: 401
        message:
          type: string
          example: Unauthorized
        errors:
          type: object
          nullable: true
      type: object
  responses:
    UnauthorizedError:
      description: 'Unauthorized - missing or invalid access token.'
      content:
        application/json:
          schema:
            type: object
            allOf:
              -
                $ref: '#/components/schemas/ApiError'
              -
                properties:
                  success:
                    type: boolean
                    example: false
                  status_code:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: 'Unauthorized. Access token is missing or invalid.'
                  errors:
                    type: object
                    example: null
                    nullable: true
                type: object
...
    NotFoundError:
      description: 'Not Found - requested resource was not found.'
      content:
        application/json:
          schema:
            type: object
            allOf:
              -
                $ref: '#/components/schemas/ApiError'
              -
                properties:
                  success:
                    type: boolean
                    example: false
                  status_code:
                    type: integer
                    example: 404
                  message:
                    type: string
                    example: 'Resource not found.'
                  errors:
                    type: object
                    example: null
                    nullable: true
                type: object
    ConflictError:
      description: 'Conflict - resource already exists or state conflict occurred.'
      content:
        application/json:
          schema:
            type: object
            allOf:
              -
                $ref: '#/components/schemas/ApiError'
              -
                properties:
                  success:
                    type: boolean
                    example: false
                  status_code:
                    type: integer
                    example: 409
                  message:
                    type: string
                    example: 'Conflict. Resource state already exists.'
                  errors:
                    type: object
                    example: null
                    nullable: true
                type: object
...
    ServerError:
      description: 'Internal Server Error - an unexpected server error occurred.'
      content:
        application/json:
          schema:
            type: object
            allOf:
              -
                $ref: '#/components/schemas/ApiError'
              -
                properties:
                  success:
                    type: boolean
                    example: false
                  status_code:
                    type: integer
                    example: 500
                  message:
                    type: string
                    example: 'Internal server error.'
                  errors:
                    type: object
                    example: null
                    nullable: true
                type: object
...
  securitySchemes:
    BearerAuth:
      type: http
      description: 'Use Sanctum access token in Authorization header as: Bearer {token}'
      bearerFormat: JWT
      scheme: bearer
...